src/requirements.txt
argparse
bokeh
datetime
elasticsearch
geoip
gower
pathlib
pywebio
requests
rich
tabulate
tldextract
pyyaml
gitpython
plotly
numpy
scipy
sklearn
pandas
pytest
